[scanner] refactor: split MCP workload queries into focused modules
Fixes #18909

Split three large workload query files into focused modules:

- pods.ts (721 lines) → pods/{list.ts (476), issues.ts (259), index.ts (14)}
- infrastructure.ts (663 lines) → infrastructure/{workloads.ts (610), logs.ts (56), index.ts (13)}
- deployments.ts (537 lines) → deployments/{queries.ts (539), cache.ts (5), index.ts (9)}

All original exports preserved via re-export barrels.
