+++ web/src/hooks/mcp/workloadQueries/deployments/cache.ts
@@ -1,0 +1,5 @@
+import { __resetDeploymentQueryCaches } from './queries'
+
+export function resetDeploymentsCache() {
+  __resetDeploymentQueryCaches()
+}
