+++ web/src/hooks/mcp/workloadQueries/deployments/index.ts
@@ -1,0 +1,9 @@
+export {
+  getDemoDeploymentIssues,
+  getDemoDeployments,
+  useDeploymentIssues,
+  useDeployments,
+} from './queries'
+export {
+  resetDeploymentsCache,
+} from './cache'
