+++ web/src/hooks/mcp/workloadQueries/infrastructure/index.ts
@@ -1,0 +1,13 @@
+export {
+  __resetInfrastructureCaches,
+  useJobs,
+  useHPAs,
+  useReplicaSets,
+  useStatefulSets,
+  useDaemonSets,
+  useCronJobs,
+} from './workloads'
+export {
+  USE_POD_LOGS_DEFAULT_TAIL,
+  usePodLogs,
+} from './logs'
